Adding filesize check

+++ gdrive_sync.rb
@@ -1,5 +1,4 @@
 require "google_drive"
-require "rb-inotify"
 require 'json'
 
 # Get the Google Drive subfolder to upload files to
@@ -7,29 +6,35 @@ config = File.read('config.json')
 googlefolder = JSON.parse(config)['subfolder']
 
 foldersync = '/files'
-#notifier = INotify::Notifier.new
 
-# Checks if a file has changed in the last 2 seconds
+# Checks if a file has changed in the last 5 seconds
 def checkforChange(filepath)
-  size = 0
-  puts File.size(filepath)
+  size1 = File.size(filepath)
   sleep 5
-  puts File.size(filepath)
+  size2 = File.size(filepath)
+  if size1 == size2
+    0
+  else
+    1
+  end
 end
 
 def uploadFile(file,foldersync,googlefolder)
   session = GoogleDrive::Session.from_config("config.json")
   basename = File.basename(file)
-  puts "\tCopying #{file}"
-  checkforChange(file)
-  exit
-  uploaded = session.upload_from_file("#{foldersync}/#{basename}", basename, convert: false)
-  unless googlefolder.nil?
-    puts "\tMoving it to #{googlefolder}" 
-    session.collection_by_title("Plex Cloud Sync").add(uploaded)
+  status = checkforChange(file)
+  if status == 0
+    puts "\tCopying #{file} to Google Drive"
+    uploaded = session.upload_from_file("#{foldersync}/#{basename}", basename, convert: false)
+    unless googlefolder.nil?
+      puts "\tMoving it to #{googlefolder}" 
+      session.collection_by_title("Plex Cloud Sync").add(uploaded)
+    end
+    session.root_collection.remove(uploaded)
+    File.delete("#{foldersync}/#{basename}")
+  else
+    puts "File hasn't arrived yet...waiting for it to finish"
   end
-  session.root_collection.remove(uploaded)
-  File.delete("#{foldersync}/#{basename}")
 end
 
 loop do
